requirements.txt
numpy~=1.26.4
torchvision~=0.20.0
scikit-learn~=1.3.0
tqdm~=4.66.5
pillow~=10.4.0